requirements.txt
## requirements.txt (otimizado)

# Core dependencies

flask>=2.0
redis>=4.0
celery>=5.0
python-dotenv>=1.0
firebase-admin>=6.0
requests>=2.31.0
pytz>=2021.1

# Web/API extras

flask-cors>=3.0
flask-limiter>=2.0
flask-socketio>=5.0
waitress>=2.1

# Asynchronous & messaging

eventlet>=0.33
websockets>=11.0
aiohttp>=3.8

# Data processing & ML

numpy<2.0
pandas>=2.0
opencv-python-headless>=4.7
torch>=2.0
transformers>=4.0
openai>=0.27
openai-whisper>=2023.7
huggingface-hub>=0.15

# Browser automation

playwright>=1.35

# Utilities

pyyaml>=6.0
user-agents>=2.2
stripe>=5.0
python-telegram-bot>=20.0
discord.py>=2.0
psutil>=5.9
schedule>=1.2
termcolor>=2.0
wave>=0.0.2
srt>=3.5
yt-dlp>=2023.12.1
proglog>=0.1.10
GPUtil>=1.4
pybind11>=2.10

# Optional multimedia

moviepy>=1.0
av>=10.0
imageio>=2.33
imageio-ffmpeg>=0.4.9

# Google APIs

google-auth>=2.0
google-auth-oauthlib>=0.7
google-auth-httplib2>=0.1
google-api-python-client>=2.0

# Development & debugging

python-dotenv>=1.0
tqdm>=4.66
typing-extensions>=4.9
reportlab>=3.6
obsws-python>=0.6

# Format & lint (opcionais)

# black

# flake8

# Endereço para pacotes específicos:

# pytube @ git+[https://github.com/pytube/pytube@a32fff39058a6f7e5e59ecd06a7467b71197ce35](https://github.com/pytube/pytube@a32fff39058a6f7e5e59ecd06a7467b71197ce35)
